@@ -1,19 +1,19 @@
 [package]
 name = "sticker-bot"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [profile.release]
 strip = true
 lto = true
 
 [dependencies]
-teloxide = { version = "0", features = ["rustls"] }
-log = "0"
+teloxide = { version = "0.17", features = ["rustls"] }
+log = "0.4"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "process"] }
-image = "0"
+image = "0.25"
 anyhow = "1"
-webp = "0"
+webp = "0.3"
 bytes = "1"
 tempfile = "3"
 pretty_env_logger = { git = "https://github.com/karin0/pretty-env-logger.git" }
